Blender module: Add prerender_count option - a dirty hack to make objects initialized properly.
Sometimes (especially when using linked groups with armatures and
complex expressions) the objects are not properly initialized
and appear correctly only on second-third render.
With this trick you can instruct blender to make some count of
pre-renders to ensure that all objects are properly initialized.

Just put a relevant option into .conf file, like this:
     blender_prerender_count=1