Allow to explicitly disable gpu rendering
This can be done in two ways:

1. By setting any value to BLENDER_DISABLE_GPU environment variable
This allows to globally for current render instance.
Such option is useful in the following setup: two render instances are running on the same computer. One uses GPU and other utilizes CPU only.

2. By setting disable_cpu=1 in any .conf file.
This allows to explicitly disable GPU rendering for particular file.