Implement two options to control dimensions of output
1. 'use_own_dimensions' - If set to '1'. then RenderChan uses dimensions of source file.
This is supported only if module can retrieve width and height of file.
See analyze() in krita.py file for sample.
This prameter can defined for particular module (format). For example it is set to '1' for Krita module.
It is can be overriden for individual files through .conf file.
Default value: '0'

When 'use_own_dimensions' is set to '1' we need some way to control the scaling of the output depending on the profile. So here goes a second option:

2. 'proxy_scale' - This parameter is applied only if 'use_own_dimensions' set to '1'.
It applies the scaling factor to the output.
In usual case this option should be set in project.conf file for individual profiles.
For example, 'draft' profile can have 'proxy_scale' = '0.5'
Default value: '1.0'

Important: If dimensions, produced by scaling, are not integer values, then no scaling is applied. This is done to avoid distortions for image proportions.