Introduce "RENDERCHAN_ENVDIR" environment variable to define location of "env" directory.
The "env" directory contains text files, which define paths to applications.
It is intended to define custom locations for apps - Blender, Synfig, etc.

See sample contents of "env" directory in "env/windows" dir of this repository.

With introduction of "RENDERCHAN_ENVDIR" environment variable, users can have various environment profiles.
For example, it is possible to have a profile which uses Blender 2.79 for one project and second profile which uses Blender 2.82 for other project.